Fetches proposition from API using a query
Source: R/proposicoes.R

Fetches information about law's projects, resolutions, provisional measures, law amendments, opinions and all the other propositions types on the Deputies' Chamber. Several parameters can be used to select and filter the final result. By default, the function returns all the proposition which were presented or had some situation change in the last 15 days.

fetch_proposicao(id = NULL, siglaUfAutor = NULL, siglaTipo = NULL,
  siglaPartidoAutor = NULL, numero = NULL, ano = NULL,
  dataApresentacaoInicio = NULL, dataApresentacaoFim = NULL,
  dataInicio = NULL, dataFim = NULL, idAutor = NULL, autor = NULL,
  codPartido = NULL, itens = NULL)

Arguments

id
Proposition's ID

siglaUfAutor
State's abbreviation of the proposition's author

siglaTipo
Proposition type (i.e., PEC, PL, PDC)

siglaPartidoAutor
Party's abbreviation of the proposition's author

numero
Proposition number

ano
Proposition year

dataApresentacaoInicio
Proposition's presentation starting date

dataApresentacaoFim
Proposition's presentation end date

dataInicio
Proposition's processing starting date

dataFim
Proposition's processing end date

idAutor
Author's ID

autor
Author's name

codPartido
Party code

itens
Items quantity. '-1' returns all the propositions which had been processed in the last 30 days

Value

Dataframe containing information about the proposition.

Details

Note that if you have the proposition's ID, it's not necessary to add any other parameter on the function call. The call to this function using the proposition's ID returns more details than using the others parameters. If you don't have the proposition's ID, the fetch_id_proposicao function may be helpful.

See also

Examples

pec241 <- fetch_proposicao(id = 2088351)
pec241 <- fetch_proposicao(siglaTipo = "PEC", numero = 241, ano = 2016)